+++ lib/dalli.rb
@@ -4,6 +4,12 @@ require 'dalli/server'
 require 'dalli/version'
 require 'dalli/options'
 
+unless String.respond_to?(:bytesize)
+  class String
+    alias_method :bytesize, :size
+  end
+end
+
 module Dalli
   # generic error
   class DalliError < RuntimeError; end