+++ lib/dalli/client.rb
@@ -185,7 +185,9 @@ module Dalli
     end
     
     def validate_key(key)
-      raise ArgumentError, "illegal character in key #{key.inspect}" if key =~ /\s/
+      raise ArgumentError, "illegal character in key #{key}" if key.respond_to?(:ascii_only?) && !key.ascii_only?
+      raise ArgumentError, "illegal character in key #{key}" if key =~ /\s/
+      raise ArgumentError, "illegal character in key #{key}" if key =~ /[\x00-\x20\x80-\xFF]/
       raise ArgumentError, "key cannot be blank" if key.nil? || key.strip.size == 0
       raise ArgumentError, "key too long #{key.inspect}" if key.length > 250
       @options[:namespace] ? "#{@options[:namespace]}:#{key}" : key